test/web-platform-tests/to-upstream/css/css-conditional/container-rule-malformed-prelude.html

<!doctype html>
<title>@container with malformed prelude is silently dropped</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">

<style id="malformed-prelude">
  @container ) { .a { color: red } }
</style>

<script>
"use strict";

test(() => {
  const { sheet } = document.getElementById("malformed-prelude");
  assert_equals(sheet.cssRules.length, 0, "Malformed @container prelude should produce no rules");
}, "@container with invalid prelude is dropped without throwing");
</script>
